+++ src/Objects/AbstractObject.php
@@ -15,7 +15,7 @@ abstract class AbstractObject extends AbstractResponse implements
 
     public function offsetSet($offset, $value)
     {
-        if (is_null($offset) || (int) $offset === $offset) {
+        if (is_null($offset) || is_int($offset)) {
            throw new \RuntimeException('Sorry, indexed arrays not allowed at the root of '.get_class($this).' objects.');
         }
 
@@ -37,11 +37,33 @@ abstract class AbstractObject extends AbstractResponse implements
         return $this;
     }
 
-    public function mergeValue(ValueInterface $item)
+    public function mergeValue($object)
     {
-        if ($item) {
-            $this->_mergeObject($item);
+        if (is_object($object)) {
+            
+            $this->_mergeObject($object);
+
+        } else if (is_array($object)) {
+
+            foreach ($object as $key => $value) {
+
+                if (is_int($key)) {
+                    $this[(int) $key] = null; // will force throw the exception above
+                }
+                
+                $key = (string) $key;
+                if (is_array($value)) {
+                    $value = new ObjectArray($value);
+                }
+
+                if (isset($this->{$key}) && is_object($value) && is_object($this->{$key})) {
+                    $this->_mergeObject($value, $this->{$key});
+                } else {
+                    $this->{$key} = $value;
+                }
+            }
         }
+        return $this;
     }
 
     public function resetValue()